fix: 修复窗口坐标 x 为负数时候选框位置偏移
windowPosition 返回值应该为有符号

+++ src/addons/wlfrontend/X11AppMonitor.cpp
@@ -67,7 +67,7 @@ pid_t X11AppMonitor::getWindowPid(xcb_window_t window)
     return *reinterpret_cast<uint32_t *>(data1.data());
 }
 
-std::tuple<uint16_t, uint16_t> X11AppMonitor::getWindowPosition(xcb_window_t window)
+std::tuple<int32_t, int32_t> X11AppMonitor::getWindowPosition(xcb_window_t window)
 {
     if (window == 0) {
         return { 0, 0 };
